refactor: 도메인의 component 어노테이션 클래스 서비스로 이동
