+++ kitchenpos-core/src/main/java/kitchenpos/ordertable/application/OrderStatusValidator.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.domain;
+package kitchenpos.ordertable.application;
 
 public interface OrderStatusValidator {
